% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBIOS.R
\name{SBIOS}
\alias{SBIOS}
\title{Scalable Bayesian Image On Scalar regression}
\usage{
SBIOS(
  img_list,
  predictor,
  grids,
  out_path,
  region_idx = NULL,
  mask_list = NULL,
  basis = NULL,
  subsample_size = 100,
  init_params = NULL,
  controls = NULL,
  n_batch = NULL,
  common_mask = NULL
)
}
\arguments{
\item{img_list}{A list of path to all image Nifti data, length(img_list) is the number of subject}

\item{predictor}{A matrix of dimension (q+1) x n}

\item{grids}{A matrix of dimension p x d, d is the dimension of the image, could be 2D or 3D}

\item{out_path}{The directory to save the coefficient beta (beta_mean) and its marginal inclusion probability (marginal_IP) as nifti file}

\item{region_idx}{A list of length num_region. Users are allowed to pre-specify
the region parcellation by putting indices for each region as each component in this list.
The dafault value is list(region1 = 1:p) with only 1 region.}

\item{mask_list}{A list of path to all mask Nifti data. Default is set to NULL, and not use any imputation.}

\item{basis}{A list object to speficy 
\itemize{
     \item Phi_Q A list object with length num_region.
    Each component represents one basis function, a matrix of dimension p_r by L_r for the r-th region.
    \item Phi_D A list object with length num_region.
    Each component represents one set of eigenvalues for one region, a vector of length L_r for the r-th region.
}}

\item{subsample_size}{An integer, the subsample size used in the SGLD algorithm, must be smaller than the total number of subjects in each batch of data.}

\item{init_params}{A list of initial parameters
\itemize{
     \item theta_beta, A vector of length L, where L is the total number of basis functions. Default value is rep(1,L).
     \item theta_gamma, A q by L matrix, default at 1
     \item theta_eta, A L by n matrix, default at 0
     \item eta, A p by n matrix, default at 0
     \item delta, A p by 1 vector, default at 1
     \item sigma_Y, default at 1
     \item sigma_beta, default at 1
     \item sigma_eta, default at 1
     \item sigma_gamma, default at 1
}}

\item{controls}{A list of control options
\itemize{
     \item lambda, thresholding parameter, default at 0.5
     \item prior_p, the prior of delta follows bernoulli(prior_p), default at 0.5
     \item n_mcmc, total number of iterations, default at 5000
     \item start_delta, at which iteration to update delta, default at 0
     \item step, initial step size, set at 1e-3
     \item burnin, from which iteration to start saving MCMC samples, default at 0
     \item thinning, interval to save MCMC samples, default at 1,
     \item interval_eta, interval to update eta, dafault at 100,
     \item start_saving_imp, from which iteration to start save the imputed outcomes, default at 1000
     \item seed, set random seed, default at sample(1:1e5,1)
     \item step_controls, A list object of component a, b, gamma, adjust the step size using: step = a*(b+iter)^gamma
}}

\item{n_batch}{number of batches}

\item{common_mask}{A vector of indices to indicate which voxels in grids(rows) to be included as the common area to perform the analysys}
}
\description{
SBIOS function is an illustrative example for working with small data sets.
For large scale imaging mediation analysis, please run SBIOSimp or SBIOS0 with detailed options and controls. 
See README.md for an example
}
